vote page

+++ console/migrate/migration/1474812248_city.php
@@ -5,7 +5,7 @@ $q = array();
 $q[] = 'CREATE TABLE `city`
         (
             `city_id` INT(11) PRIMARY KEY AUTO_INCREMENT,
-            `city_country_id` INT(11) DEFAULT 0,
+            `city_country_id` SMALLINT(3) DEFAULT 0,
             `city_name` VARCHAR(255) NOT NULL
         );';
 $q[] = 'CREATE INDEX `city_country_id` ON `city` (`city_country_id`);';
