+++ console/migrate/migration/1474817009_stadium.php
@@ -5,7 +5,7 @@ $q = array();
 $q[] = 'CREATE TABLE `stadium`
         (
             `stadium_id` INT(11) PRIMARY KEY AUTO_INCREMENT,
-            `stadium_capacity` INT(11) DEFAULT 100,
+            `stadium_capacity` SMALLINT(5) DEFAULT 100,
             `stadium_city_id` INT(11) DEFAULT 0,
             `stadium_name` VARCHAR(255) NOT NULL
         );';
