+++ console/migrate/migration/1475586415_sex.php
@@ -4,7 +4,7 @@ $q = array();
 
 $q[] = 'CREATE TABLE `sex`
         (
-            `sex_id` INT(11) PRIMARY KEY AUTO_INCREMENT,
+            `sex_id` TINYINT(1) PRIMARY KEY AUTO_INCREMENT,
             `sex_name` VARCHAR(255) NOT NULL
         );';
 $q[] = "INSERT INTO `sex` (`sex_name`)
