+++ console/migrate/migration/1475586429_style.php
@@ -4,6 +4,6 @@ $q = array();
 
 $q[] = 'CREATE TABLE `style`
         (
-            `style_id` INT(11) PRIMARY KEY AUTO_INCREMENT,
+            `style_id` TINYINT(1) PRIMARY KEY AUTO_INCREMENT,
             `style_name` VARCHAR(255) NOT NULL
         );';
