+++ console/migrate/migration/1475586464_special.php
@@ -4,6 +4,6 @@ $q = array();
 
 $q[] = 'CREATE TABLE `special`
         (
-            `special_id` INT(11) PRIMARY KEY AUTO_INCREMENT,
+            `special_id` TINYINT(2) PRIMARY KEY AUTO_INCREMENT,
             `special_name` VARCHAR(255) NOT NULL
         );';
