+++ console/migrate/migration/1475587849_playerposition.php
@@ -5,7 +5,7 @@ $q = array();
 $q[] = 'CREATE TABLE `playerposition`
         (
             `playerposition_player_id` INT(11) DEFAULT 0,
-            `playerposition_position_id` INT(11) DEFAULT 0
+            `playerposition_position_id` TINYINT(1) DEFAULT 0
         );';
 $q[] = 'CREATE INDEX `playerposition_player_id` ON `playerposition` (`playerposition_player_id`);';
 $q[] = 'CREATE INDEX `playerposition_position_id` ON `playerposition` (`playerposition_position_id`);';
