+++ console/migrate/migration/1475670698_season.php
@@ -4,7 +4,7 @@ $q = array();
 
 $q[] = 'CREATE TABLE `season`
         (
-            `season_id` INT(11) PRIMARY KEY AUTO_INCREMENT
+            `season_id` SMALLINT(5) PRIMARY KEY AUTO_INCREMENT
         );';
 $q[] = 'INSERT INTO `season`
         SET `season_id`=NULL';
