+++ console/migrate/migration/1477389998_rule.php
@@ -4,9 +4,9 @@ $q = array();
 
 $q[] = 'CREATE TABLE `rule`
         (
-            `rule_id` INT(11) PRIMARY KEY AUTO_INCREMENT,
+            `rule_id` TINYINT(3) PRIMARY KEY AUTO_INCREMENT,
             `rule_date` INT(11) DEFAULT 0,
-            `rule_order` INT(11) DEFAULT 0,
+            `rule_order` TINYINT(3) DEFAULT 0,
             `rule_text` TEXT NOT NULL,
             `rule_title` VARCHAR(255) NOT NULL
         );';
